+++ js/utils.js
@@ -1,0 +1,41 @@
+//генерирует случайное число в диапазоне
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+//генерирует случайное число в диапазоне без повторений
+const createRandomIdFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      console.error('Перебраны все числа из диапазона от ' + min + ' до ' + max);
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+//генерирует id без повторений без ограничения диапазона
+const createIdGenerator = () => {
+  let numberId = 0;
+
+  return () => {
+    numberId += 1;
+    return numberId;
+  };
+};
+
+const generateRandomId = createIdGenerator();
+
+export {getRandomInteger, getRandomArrayElement, generateRandomId};
